fixing subscription controller

+++ src/controllers/comment.controller.js
@@ -2,19 +2,59 @@ import { ApiError } from "../utils/ApiError.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import commentService from "../services/comment.service.js";
+import videoService from '../services/video.service.js';
+import { Video } from "../models/video.model.js"
 
-const getComment = asyncHandler( async(req,res) => {
+const getComments = asyncHandler( async(req,res) => {
+  const {videoId} = req.params
 
+  const comments = await Video.aggregate([
+    {
+      $match:{
+        _id: new mongoose.Types.ObjectId(videoId)
+      }
+    },
+    {
+      $lookup:{
+        from:'comments',
+        localField:'comments',
+        foreignField:'_id',
+        as:'comments'
+      }
+    },
+    {
+      $unwind: "$comments"
+    },
+    {
+      $project:{
+        _id:0,
+        comments:{
+          content:"$comments.content",
+          owner:"$comments.owner"
+        }
+      }
+    }
+  ])
+
+  return res.status(200).json(
+    new ApiResponse(200,comments,'Comment Fetched!')
+  )
 })
 
 const addComment = asyncHandler( async(req,res) => {
   const {content} = req.body
   const {videoId} = req.params
 
+  const video = await videoService.findVideoById(videoId)
+
+  if(!video){
+    throw new ApiError(404,'Video not found!')
+  }
+
   const commentData = {
     content,
     video:videoId,
-    owner:req.user?._Id
+    owner:req.user?._id
   }
 
   const createdComment = await commentService.createComment(commentData)
@@ -22,6 +62,9 @@ const addComment = asyncHandler( async(req,res) => {
   if(!createdComment){
     throw new ApiError(501, "Comment couldn't be added!");
   }
+
+  video.comments.push(createdComment?._id)
+  await video.save()
 
   return res.status(201).json(
     new ApiResponse(201,createdComment,'comment added!')
@@ -37,7 +80,10 @@ const editComment = asyncHandler( async(req,res) => {
   if(!comment){
     throw new ApiError(404,'Comment not found!')
   }
-  // only commentOwner can edit thier comments
+  
+  if(!(req.user?._id.toString()===comment.owner.toString())){
+    return new ApiError(401,'Unauthorized request!')
+  }
 
   const editedComment = await commentService.editComment(
     commentId,
@@ -49,7 +95,7 @@ const editComment = asyncHandler( async(req,res) => {
   )
 
   if(!editedComment){
-    throw new ApiError(501, "Comment couldn't be updated!");
+    throw new ApiError(501,"Comment couldn't be updated!");
   }
     
   return res.status(201).json(
@@ -65,18 +111,22 @@ const deleteComment = asyncHandler( async(req,res) => {
   if(!comment){
     throw new ApiError(404,'Comment not found!')
   }
-  // only commentOwner can delete thier comments
-  
+
+  if(!(req.user?._id.toString()===comment.owner.toString())){
+    return new ApiError(401,'Unauthorized request!')
+  }
+
   const isDeleted = await commentService.deleteComment(commentId)
 
   if(!isDeleted){
     throw new ApiError(401,{},"Comment couldn't be deleted!")
   }
-  // delete the commentLikes also
+  
+  await commentService.deleteLikes(commentId)
 
   res.status(201).json(
     new ApiResponse(201,{},'Comment deleted successfully!')
   )
 })
 
-export {getComment, addComment, editComment, deleteComment}
+export {getComments, addComment, editComment, deleteComment}
